requirements.txt
arrow==1.2.3
attrs==23.1.0
Automat==22.10.0
boto==2.49.0
boto3==1.26.132
botocore==1.29.132
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
elastic-transport==8.4.0
elasticsearch==7.13.4
filelock==3.12.0
flake8==6.0.0
fqdn==1.5.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
parsel==1.8.1
premailer==3.10.0
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.7
pyflakes==3.0.1
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
queuelib==1.6.2
requests==2.30.0
requests-file==1.5.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3987==1.3.8
s3transfer==0.6.1
schematics==2.1.1
scrapinghub==2.4.0
Scrapy==2.9.0
ScrapyElasticSearch==0.9.2
sentry-sdk==1.22.2
service-identity==21.1.0
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
spidermon==1.18.0
text-unidecode==1.3
tldextract==3.4.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
w3lib==2.1.1
webcolors==1.13
zope.interface==6.0
